services/NorthernFiles/auth/requirements.txt
APIFlask==1.3.1
apispec==6.3.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
Deprecated==1.2.14
Flask==2.3.2
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
marshmallow==3.19.0
mysqlclient==2.1.1
packaging==23.1
pycparser==2.21
pycryptodome==3.18.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
webargs==8.2.0
Werkzeug==2.3.6
wrapt==1.15.0
